Minor additions and changes
Allow for defense stat to be selected with Armor. Might not be the right place for it, and if it isn't, I might change it later.
Made the roll dialogue a bit more pleasant and adaptive (not prompting for target stat if no target is selected, show less fields if not required, saved details for rolls)
Max/min of 19 or 1 as results and other minor changes for rolls
Made rolls show character name instead of user to ease the understanding on who we are rolling for when using familiars/etc.